Shrinked translation array, fixed compilation with MSVC10
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/trunk@366 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ quadtree.cpp
@@ -1185,8 +1185,8 @@ static int get_root_level (int nx, int nz)
 {
 	int xlev, zlev;
 
-	xlev = (int) (log(nx) / log (2.0));
-	zlev = (int) (log(nz) / log (2.0));
+	xlev = (int) (log(static_cast<double>(nx)) / log (2.0));
+	zlev = (int) (log(static_cast<double>(nz)) / log (2.0));
 	if ((nx >> xlev) << xlev == nx) {
 	} else nx += 1;
